+++ 301/CO_HBconsumer.c
@@ -49,6 +49,13 @@ static void CO_HBcons_receive(void *object, void *msg){
         HBconsNode->NMTstate = (CO_NMT_internalState_t)data[0];
         CO_FLAG_SET(HBconsNode->CANrxNew);
     }
+
+#if CO_CONFIG_HB_CONS & CO_CONFIG_FLAG_CANRX_CALLBACK
+    /* Optional signal to RTOS, which can resume task, which handles HBcons. */
+    if(HBconsNode->pFunctSignalCanRx != NULL) {
+        HBconsNode->pFunctSignalCanRx(HBconsNode->functSignalObjectCanRx);
+    }
+#endif
 }
 
 
@@ -116,11 +123,22 @@ CO_ReturnError_t CO_HBconsumer_init(
     HBcons->NMTisPreOrOperationalPrev = false;
     HBcons->CANdevRx = CANdevRx;
     HBcons->CANdevRxIdxStart = CANdevRxIdxStart;
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_CHANGE_CALLBACK
+    HBcons->pFunctSignalNmtChanged = NULL;
+#endif
 
     for(i=0; i<HBcons->numberOfMonitoredNodes; i++) {
         uint8_t nodeId = (HBcons->HBconsTime[i] >> 16U) & 0xFFU;
         uint16_t time = HBcons->HBconsTime[i] & 0xFFFFU;
         ret = CO_HBconsumer_initEntry(HBcons, i, nodeId, time);
+#if CO_CONFIG_HB_CONS & CO_CONFIG_FLAG_CANRX_CALLBACK
+            HBcons->monitoredNodes[i].pFunctSignalCanRx = NULL;
+#endif
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_MULTI_CALLBACK
+            HBcons->monitoredNodes[i].pFunctSignalHbStarted = NULL;
+            HBcons->monitoredNodes[i].pFunctSignalTimeout = NULL;
+            HBcons->monitoredNodes[i].pFunctSignalRemoteReset = NULL;
+#endif
     }
 
     /* Configure Object dictionary entry at index 0x1016 */
@@ -165,7 +183,7 @@ CO_ReturnError_t CO_HBconsumer_initEntry(
         monitoredNode->nodeId = nodeId;
         monitoredNode->time_us = (int32_t)consumerTime_ms * 1000;
         monitoredNode->NMTstate = CO_NMT_INITIALIZING;
-#if CO_CONFIG_HB_CONS_CALLBACKS & 1
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_CHANGE_CALLBACK
         monitoredNode->NMTstatePrev = CO_NMT_INITIALIZING;
 #endif
         CO_FLAG_CLEAR(monitoredNode->CANrxNew);
@@ -196,7 +214,25 @@ CO_ReturnError_t CO_HBconsumer_initEntry(
 }
 
 
-#if CO_CONFIG_HB_CONS_CALLBACKS & 1
+#if CO_CONFIG_HB_CONS & CO_CONFIG_FLAG_CANRX_CALLBACK
+/******************************************************************************/
+void CO_HBconsumer_initCallback(
+        CO_HBconsumer_t        *HBcons,
+        void                   *object,
+        void                  (*pFunctSignal)(void *object))
+{
+    if (HBcons != NULL) {
+        uint8_t i;
+        for(i=0; i<HBcons->numberOfMonitoredNodes; i++) {
+            HBcons->monitoredNodes[i].pFunctSignalCanRx = pFunctSignal;
+            HBcons->monitoredNodes[i].functSignalObjectCanRx = object;
+        }
+    }
+}
+#endif
+
+
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_CHANGE_CALLBACK
 /******************************************************************************/
 void CO_HBconsumer_initCallbackNmtChanged(
         CO_HBconsumer_t        *HBcons,
@@ -212,10 +248,10 @@ void CO_HBconsumer_initCallbackNmtChanged(
     HBcons->pFunctSignalNmtChanged = pFunctSignal;
     HBcons->pFunctSignalObjectNmtChanged = object;
 }
-#endif /* CO_CONFIG_HB_CONS_CALLBACKS & 1 */
+#endif
 
 
-#if CO_CONFIG_HB_CONS_CALLBACKS & 2
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_MULTI_CALLBACK
 /******************************************************************************/
 void CO_HBconsumer_initCallbackHeartbeatStarted(
     CO_HBconsumer_t        *HBcons,
@@ -271,7 +307,7 @@ void CO_HBconsumer_initCallbackRemoteReset(
     monitoredNode->pFunctSignalRemoteReset = pFunctSignal;
     monitoredNode->functSignalObjectRemoteReset = object;
 }
-#endif /* CO_CONFIG_HB_CONS_CALLBACKS & 2 */
+#endif /* CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_MULTI_CALLBACK */
 
 
 /******************************************************************************/
@@ -298,7 +334,7 @@ void CO_HBconsumer_process(
             if (CO_FLAG_READ(monitoredNode->CANrxNew)) {
                 if (monitoredNode->NMTstate == CO_NMT_INITIALIZING) {
                     /* bootup message*/
-#if CO_CONFIG_HB_CONS_CALLBACKS & 2
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_MULTI_CALLBACK
                     if (monitoredNode->pFunctSignalRemoteReset != NULL) {
                         monitoredNode->pFunctSignalRemoteReset(
                             monitoredNode->nodeId, i,
@@ -315,7 +351,7 @@ void CO_HBconsumer_process(
                 }
                 else {
                     /* heartbeat message */
-#if CO_CONFIG_HB_CONS_CALLBACKS & 2
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_MULTI_CALLBACK
                     if (monitoredNode->HBstate != CO_HBconsumer_ACTIVE &&
                         monitoredNode->pFunctSignalHbStarted != NULL) {
                         monitoredNode->pFunctSignalHbStarted(
@@ -337,7 +373,7 @@ void CO_HBconsumer_process(
 
                 if (monitoredNode->timeoutTimer >= monitoredNode->time_us) {
                     /* timeout expired */
-#if CO_CONFIG_HB_CONS_CALLBACKS & 2
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_MULTI_CALLBACK
                     if (monitoredNode->pFunctSignalTimeout!=NULL) {
                         monitoredNode->pFunctSignalTimeout(
                             monitoredNode->nodeId, i,
@@ -366,7 +402,7 @@ void CO_HBconsumer_process(
             if (monitoredNode->NMTstate != CO_NMT_OPERATIONAL) {
                 allMonitoredOperationalCurrent = CO_NMT_INITIALIZING;
             }
-#if CO_CONFIG_HB_CONS_CALLBACKS & 1
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_CHANGE_CALLBACK
             /* Verify, if NMT state of monitored node changed */
             if(monitoredNode->NMTstate != monitoredNode->NMTstatePrev) {
                 if (HBcons->pFunctSignalNmtChanged != NULL) {
@@ -384,7 +420,7 @@ void CO_HBconsumer_process(
         /* (pre)operational state changed, clear variables */
         for(i=0; i<HBcons->numberOfMonitoredNodes; i++) {
             monitoredNode->NMTstate = CO_NMT_INITIALIZING;
-#if CO_CONFIG_HB_CONS_CALLBACKS & 1
+#if CO_CONFIG_HB_CONS & CO_CONFIG_HB_CONS_CHANGE_CALLBACK
             monitoredNode->NMTstatePrev = CO_NMT_INITIALIZING;
 #endif
             CO_FLAG_CLEAR(monitoredNode->CANrxNew);